+++ tests/Query/UpdateQueryTest.php
@@ -29,7 +29,7 @@ class UpdateQueryTest extends PHPUnit_Framework_TestCase
         $this->assertEquals($query, $query->values(['test1' => 1, 'test2' => 2]));
         $this->assertEquals($query, $query->values(['test3' => 3]));
         $this->assertInstanceOf('JAQB\Statement\SetStatement', $query->getSet());
-        $this->assertEquals(['test1' => 1, 'test2' => 2, 'test3' => 3], $query->getSet()->getValues());
+        $this->assertEquals(['test1' => 1, 'test2' => 2, 'test3' => 3], $query->getSet()->getSetValues());
     }
 
     public function testWhere()