+++ CHANGELOG.md
@@ -7,6 +7,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Support for IN statement as a WHERE condition when passing in an array value.
 - Select queries can now be unioned together (using SQL UNION).
 
+### Changed
+- SetStatement `getValues()` now returns parameterized values and `getSetValues()` returns the key-value map of values being set.
+
 ### Fixed
  - Where statement writes `null` values in conditions as `IS NULL`
  - Deep cloning now works